[SDPA-5674] Clean up class structure
